Update ssh keys, rearrange site keywords and remove dev.to social link
The ssh keys in the ssh.txt file were modified to improve security. The site keywords in the config.toml file were rearranged to enhance SEO. Lastly, the dev.to social link was removed from the user's profile to maintain up-to-date information. The path and url in the .gitmodules have been updated for accurate submodule reference.